requirements.txt
flask
flask_sqlalchemy
flask_login
psycopg2-binary
flask-wtf